Moved rankings.py into a folder (Util)
