+++ StudentTrackingSystem/StudentTrackingSystemApp/Util/rankings.py
